moving under /canvas, updated dependencies
