+++ app/src/app/canvas/page.tsx
@@ -1,4 +1,4 @@
-import CanvasEditor from "../_components/editor/editor";
+import CanvasEditor from "../_components/canvas/editor";
 
 export default function page(){
     return(<CanvasEditor />);
